2/pyproject.toml
[project]
name = "2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.7.1.1",
    "openai-agents>=0.2.9",
    "python-dotenv>=1.1.1",
]
